// src/api/generated/models/purchaseSelectionSummaryStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type PurchaseSelectionSummaryStatus = typeof PurchaseSelectionSummaryStatus[keyof typeof PurchaseSelectionSummaryStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PurchaseSelectionSummaryStatus = {
  DRAFT: 'DRAFT',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
} as const;